chore: move zod logic packages/api
