+++ packages/api/src/index.ts
@@ -2,6 +2,7 @@
 export * from './mcp/manager';
 export * from './mcp/oauth';
 export * from './mcp/auth';
+export * from './mcp/zod';
 /* Utilities */
 export * from './mcp/utils';
 export * from './utils';
