+++ packages/api/src/mcp/types/index.ts
@@ -7,8 +7,9 @@ import {
   WebSocketOptionsSchema,
   StreamableHTTPOptionsSchema,
 } from 'librechat-data-provider';
-import type { JsonSchemaType, TPlugin } from 'librechat-data-provider';
 import type * as t from '@modelcontextprotocol/sdk/types.js';
+import type { TPlugin } from 'librechat-data-provider';
+import type { JsonSchemaType } from '~/types/zod';
 
 export type StdioOptions = z.infer<typeof StdioOptionsSchema>;
 export type WebSocketOptions = z.infer<typeof WebSocketOptionsSchema>;
