+++ packages/api/src/mcp/zod.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 // zod.spec.ts
 import { z } from 'zod';
-import type { JsonSchemaType } from './zod';
+import type { JsonSchemaType } from '~/types';
 import { resolveJsonSchemaRefs, convertJsonSchemaToZod } from './zod';
 
 describe('convertJsonSchemaToZod', () => {
