+++ packages/api/src/types/index.ts
@@ -4,3 +4,4 @@ export * from './google';
 export * from './mistral';
 export * from './openai';
 export * from './run';
+export * from './zod';
