+++ packages/data-provider/src/index.ts
@@ -8,7 +8,6 @@ export * from './artifacts';
 /* schema helpers  */
 export * from './parsers';
 export * from './ocr';
-export * from './zod';
 /* custom/dynamic configurations  */
 export * from './generate';
 export * from './models';
